requirements.txt
PyQt5
requests
Pillow
